requirements.txt
connexion >= 2.6.0
connexion[swagger-ui] >= 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
pymongo~=4.3.3

python-dotenv~=0.21.1
Flask~=2.2.3
six~=1.16.0